0.0.6 new endponts and dao

+++ src/main/java/com/inz/inz/seciurity/Resource/UserAuthResoruce.java
@@ -19,7 +19,7 @@ public class UserAuthResoruce {
 
     private Boolean isBanned;
 
-    private Integer userRating;
+    private Double userRating;
 
     private List<ReportLight> reports;
 }
